Replace the 256-slot tick ring with vanilla's trigger-tick ordering
Scheduled ticks were held in a fixed ring of MAX_TICK_DELAY = 1 << 8 slots,
indexed by (offset + delay) % 256, with delay stored as a u8. Any delay past
255 ticks was therefore unrepresentable, and widening the type alone would
have been worse than the cap: 5000 % 256 = 136, so a long delay would have
fired early rather than late.

Vanilla has no ring. ScheduledTick carries an absolute triggerTick and
LevelChunkTicks drains in DRAIN_ORDER (trigger tick, then priority, then
sub-tick order) -- see ScheduledTick.java and SavedTick.java. This ports that
shape: a BTreeMap keyed by absolute trigger tick, and a monotonic per-chunk
current_tick.

Because a drain removes exactly one trigger-tick bucket, trigger time stays
implicit within a bucket, so OrderedTick's (priority, sub_tick_order)
comparator is still the correct intra-tick order and the redstone determinism
work is untouched.

The save format already matched vanilla (SavedTick stores a relative delay as
an int under key t), so worlds are unaffected; only the in-memory structure
and the delay type change. Negative saved delays are clamped rather than
wrapped.

Two blocks were merged earlier this session with their delays clamped to the
old ceiling and are now restored to vanilla values: frogspawn hatching
(3600-12000 ticks, FrogspawnBlock.java:64-66) and dried ghast hydration
(5000 ticks, DriedGhastBlock.java:43).

+++ crates/pumpkin-world/src/level.rs
@@ -995,7 +995,7 @@ impl Level {
         &self,
         block: &Block,
         block_pos: BlockPos,
-        delay: u8,
+        delay: u32,
         priority: TickPriority,
     ) {
         let tick_order = self.schedule_tick_counts.fetch_add(1, Ordering::Relaxed);
@@ -1022,7 +1022,7 @@ impl Level {
         &self,
         fluid: &Fluid,
         block_pos: BlockPos,
-        delay: u8,
+        delay: u32,
         priority: TickPriority,
     ) {
         let tick_order = self.schedule_tick_counts.fetch_add(1, Ordering::Relaxed);
