+++ crates/pumpkin/src/block/blocks/fire/fire.rs
@@ -196,7 +196,7 @@ impl BlockBehaviour for FireBlock {
         args.world.schedule_block_tick(
             args.block,
             *args.position,
-            Self::get_fire_tick_delay() as u8,
+            Self::get_fire_tick_delay() as u32,
             TickPriority::Normal,
         );
     }
@@ -242,7 +242,7 @@ impl BlockBehaviour for FireBlock {
         world.schedule_block_tick(
             block,
             *pos,
-            Self::get_fire_tick_delay() as u8,
+            Self::get_fire_tick_delay() as u32,
             TickPriority::Normal,
         );
