+++ crates/pumpkin/src/block/blocks/light.rs
@@ -56,7 +56,7 @@ impl BlockBehaviour for LightBlock {
             args.world.schedule_fluid_tick(
                 &Fluid::WATER,
                 *args.position,
-                Fluid::WATER.flow_speed as u8,
+                Fluid::WATER.flow_speed as u32,
                 TickPriority::Normal,
             );
         }
