+++ crates/pumpkin/src/block/blocks/redstone/pressure_plate/mod.rs
@@ -139,7 +139,7 @@ pub(crate) trait PressurePlate {
 
     fn calculate_redstone_output(&self, world: &World, block: &Block, pos: &BlockPos) -> u8;
 
-    fn tick_rate(&self) -> u8 {
+    fn tick_rate(&self) -> u32 {
         20
     }
 }
