+++ crates/pumpkin/src/block/blocks/redstone/pressure_plate/weighted.rs
@@ -106,7 +106,7 @@ impl PressurePlate for WeightedPressurePlateBlock {
         props.to_state_id(block)
     }
 
-    fn tick_rate(&self) -> u8 {
+    fn tick_rate(&self) -> u32 {
         10
     }
 }
