+++ crates/pumpkin/src/block/blocks/scaffolding.rs
@@ -66,7 +66,7 @@ impl BlockBehaviour for ScaffoldingBlock {
             args.world.schedule_fluid_tick(
                 &Fluid::WATER,
                 *args.position,
-                Fluid::WATER.flow_speed as u8,
+                Fluid::WATER.flow_speed as u32,
                 TickPriority::Normal,
             );
         }
