+++ crates/pumpkin/src/block/blocks/signs.rs
@@ -480,7 +480,7 @@ impl BlockBehaviour for SignBlock {
             args.world.schedule_fluid_tick(
                 &Fluid::WATER,
                 *args.position,
-                Fluid::WATER.flow_speed as u8,
+                Fluid::WATER.flow_speed as u32,
                 TickPriority::Normal,
             );
         }
