+++ crates/pumpkin/src/block/blocks/weathering_copper.rs
@@ -1112,7 +1112,7 @@ impl BlockBehaviour for WaxedCopperGrateBlock {
             args.world.schedule_fluid_tick(
                 &Fluid::WATER,
                 *args.position,
-                Fluid::WATER.flow_speed as u8,
+                Fluid::WATER.flow_speed as u32,
                 TickPriority::Normal,
             );
         }
