+++ crates/pumpkin/src/world/mod.rs
@@ -6132,7 +6132,7 @@ impl World {
         &self,
         block: &Block,
         block_pos: BlockPos,
-        delay: u8,
+        delay: u32,
         priority: TickPriority,
     ) {
         self.level
@@ -6143,7 +6143,7 @@ impl World {
         &self,
         fluid: &Fluid,
         block_pos: BlockPos,
-        delay: u8,
+        delay: u32,
         priority: TickPriority,
     ) {
         self.level
